requirements.txt
streamlit==1.17.0
pandas==1.4.2
openpyxl==3.0.10
google-generative-ai==0.1.1
